requirements.txt
Django==4.2.4
djangorestframework==3.14.0
python-decouple
django-filter
uvicorn 
gunicorn
mysqlclient